--- src/main/java/ru/sbt/mipt/oop/Light.java
@@ -1,23 +1,0 @@
-package ru.sbt.mipt.oop;
-
-public class Light {
-    private boolean isOn;
-    private final String id;
-
-    public Light(String id, boolean isOn) {
-        this.id = id;
-        this.isOn = isOn;
-    }
-
-    public boolean isOn() {
-        return isOn;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setOn(boolean on) {
-        isOn = on;
-    }
-}
