--- src/main/java/ru/sbt/mipt/oop/SensorEventType.java
@@ -1,5 +1,0 @@
-package ru.sbt.mipt.oop;
-
-public enum SensorEventType {
-    LIGHT_ON, LIGHT_OFF, DOOR_OPEN, DOOR_CLOSED
-}
