--- src/main/java/ru/sbt/mipt/oop/SmartHome.java
@@ -1,24 +1,0 @@
-package ru.sbt.mipt.oop;
-
-import java.util.ArrayList;
-import java.util.Collection;
-
-public class SmartHome {
-    Collection<Room> rooms;
-
-    public SmartHome() {
-        rooms = new ArrayList<>();
-    }
-
-    public SmartHome(Collection<Room> rooms) {
-        this.rooms = rooms;
-    }
-
-    public void addRoom(Room room) {
-        rooms.add(room);
-    }
-
-    public Collection<Room> getRooms() {
-        return rooms;
-    }
-}
