+++ src/main/java/ru/sbt/mipt/oop/commands/SensorCommand.java
