+++ src/main/java/ru/sbt/mipt/oop/commands/SensorCommandType.java
@@ -1,4 +1,4 @@
-package ru.sbt.mipt.oop;
+package ru.sbt.mipt.oop.commands;
 
 public enum CommandType {
     LIGHT_OFF
