+++ src/main/java/ru/sbt/mipt/oop/events/SensorEvent.java
@@ -1,5 +1,7 @@
 package ru.sbt.mipt.oop;
 
+import static ru.sbt.mipt.oop.SensorEventType.*;
+
 public class SensorEvent {
     private final SensorEventType type;
     private final String objectId;
@@ -15,6 +17,19 @@ public class SensorEvent {
 
     public String getObjectId() {
         return objectId;
+    }
+
+    public boolean isLightEvent() {
+        return (type == LIGHT_ON || type == LIGHT_OFF);
+    }
+
+    public boolean isDoorEvent() {
+        return (type == DOOR_OPEN || type == DOOR_CLOSED
+                || type == DOOR_CLOSED_IN_HALL);
+    }
+
+    public boolean isDoorClosedInHallEvent() {
+        return (type == DOOR_CLOSED_IN_HALL);
     }
 
     @Override
