+++ src/main/java/ru/sbt/mipt/oop/events/SensorEventType.java
@@ -1,0 +1,5 @@
+package ru.sbt.mipt.oop;
+
+public enum SensorEventType {
+    LIGHT_ON, LIGHT_OFF, DOOR_OPEN, DOOR_CLOSED, DOOR_CLOSED_IN_HALL
+}
